She walked to the kitchen, silently. There was no one to disturb, but the dark house made her silence herself. There was no moon, and she did not turn on a light; giuded by the microwave clock, she reached the refrigerator. It hummed.
I'm going in to work late today because we're having a lunch meeting. I'm a little depressed, because of that or because I pick my nose, or both.

Code that is specific is a machine for pulling the pits from
two-day-old yellow cherries. Code that is general is a hungry 5-year-old boy, collapsing at any provocation.

Commit message: asking a 5-year-old to pit cherries.