She held a scream in her stomach. It ran around, searching for a way out.
I really wish I knew God. I mean, I'm an atheist, and my expectation
of that changing is low, but I don't yearn any less for the comfort of
knowing that all is forgiven, that I could be granted eternal life,
that someone's got a plan and I'm a part of it. To discover that would
be the greatest relief of my life. When I'm depressed... I guess
imagine blind faith like some people contemplate suicide.

This should be the last big ugly unsemantic commit! Things are
arranged more sensibly; they are by no means perfect, but I think I
can work in small chinks without breaking everything. All of the
component are playing together nicely. The next step is probably to
replace the parser, but there are still some discussions to have
before I can do that; the other thing that is necessary is to write a
real garbage collector, which is a task I mortally dread. I may write
some about what oyster is supposed to be.

Commit message: The Screaming God mortally dreads commitment.